requirements.txt
requests
flask
flask-limiter
google.generativeai
Pillow
urllib3
rembg[gpu,cli]